+++ src/utils/genereateColor.tsx
@@ -1,0 +1,1 @@
+export {};
